rename: decompose_util
